fix: Reset feedback when skipping submission (#59)
This commit aims to prevent unintended feedback from appearing for the
next user in the same app session. Previously, if the previous user gave
a rating but skipped submitting feedback, the feedback would still
appear for the next user. This could lead to the next user accidentally
pressing the "Send feedback" button, as it was the primary button
displayed.

Closes #58